infrastructure/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-firestore==2.14.0
google-cloud-pubsub==2.19.0
python-multipart==0.0.6
